- Separate admFilter in adm-ambassadors and adm-sessions2 - Removed dynamic options in filters (it was bugged)
