+++ imports/ui/stylesheets/adm.scss
@@ -147,6 +147,10 @@ $grey: #aaa;
 
 }
 
+.results-count {
+  padding: 10px 0;
+}
+
 .table-hover > tbody > tr:hover {
   background-color: #333;
 }
